+++ test/color_test.exs
@@ -2,7 +2,6 @@ defmodule Image.ColorTest do
   use ExUnit.Case, async: true
 
   test "Sort colors" do
-    assert [[1, 5, 1], [1, 3, 2]] = Image.Color.sort([[1,5,1], [1,3,2]])
-
+    assert [[1, 5, 1], [1, 3, 2]] = Image.Color.sort([[1, 5, 1], [1, 3, 2]])
   end
 end